# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TON Site Builder API",
    "description": "Backend Telegram Mini App для создания и публикации сайтов в сети TON",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/domains/check": {
      "get": {
        "tags": [
          "domains"
        ],
        "summary": "Check Domain",
        "operationId": "check_domain_api_domains_check_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 126,
              "title": "Name"
            }
          },
          {
            "name": "tld",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "default": "ton",
              "title": "Tld"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/domains/deploy-zone": {
      "post": {
        "tags": [
          "domains"
        ],
        "summary": "Deploy Zone",
        "description": "Готовит транзакцию деплоя зоны — подписывает её кошелёк пользователя.",
        "operationId": "deploy_zone_api_domains_deploy_zone_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployZoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployZoneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/domains/confirm": {
      "post": {
        "tags": [
          "domains"
        ],
        "summary": "Confirm Domain",
        "description": "Фронт сообщает, что транзакция отправлена; проверяем состояние домена в сети.\n\nФронту на слово не верим: домен считается привязанным только если доменный\nсервис подтверждает, что он больше не свободен и принадлежит кошельку.",
        "operationId": "confirm_domain_api_domains_confirm_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeployZoneRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "format": "uuid",
            "title": "Site Id"
          },
          "domain": {
            "type": "string",
            "maxLength": 126,
            "minLength": 1,
            "title": "Domain"
          },
          "tld": {
            "type": "string",
            "title": "Tld",
            "default": "ton"
          },
          "mode": {
            "type": "string",
            "enum": [
              "proxy",
              "sbt"
            ],
            "title": "Mode",
            "default": "proxy"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "domain"
        ],
        "title": "DeployZoneRequest"
      },
      "DeployZoneResponse": {
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TonConnectTransaction"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          }
        },
        "type": "object",
        "required": [
          "transaction"
        ],
        "title": "DeployZoneResponse"
      },
...
      "DomainCheckResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "dns_item_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dns Item Address"
          },
          "collection_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Address"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          }
        },
        "type": "object",
        "required": [
          "available",
          "status",
          "domain"
        ],
        "title": "DomainCheckResponse"
      },
      "DomainConfirmRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "format": "uuid",
            "title": "Site Id"
          },
          "tx_hash": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Tx Hash"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "tx_hash"
        ],
        "title": "DomainConfirmRequest"
      },
      "DomainConfirmResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DomainConfirmResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TonConnectMessage": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "stateInit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stateinit"
          }
        },
        "type": "object",
        "required": [
          "address",
          "amount"
        ],
        "title": "TonConnectMessage"
      },
      "TonConnectTransaction": {
        "properties": {
          "validUntil": {
            "type": "integer",
            "title": "Validuntil"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/TonConnectMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "network": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network"
          },
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          }
        },
        "type": "object",
        "required": [
          "validUntil",
          "messages"
        ],
        "title": "TonConnectTransaction"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}